removed objectiveScriptableObject script
all of it's responsibilites are now handled by objectiveSequence
