--- Assets/Scripts/ObjectiveSystem/ObjectiveScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dedabe918f7e5e34ea3b3753fac36e20
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
